requirements.txt
absl-py==2.2.0
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.1.0
dill==0.3.8
evaluate==0.4.3
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
packaging==24.2
pandas==1.3.5
pillow==10.2.0
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
